Added SendMsgTest. Refactored code.| #1631
