+++ FlowCrypt/build.gradle
@@ -348,6 +348,7 @@ dependencies {
     androidTestImplementation "androidx.arch.core:core-testing:2.1.0"
     androidTestImplementation "androidx.test:core-ktx:1.4.0"
     androidTestImplementation 'androidx.test.ext:junit-ktx:1.1.3'
+    androidTestImplementation 'androidx.work:work-testing:2.7.1'
     androidTestImplementation "com.squareup.okhttp3:mockwebserver:${rootProject.ext.okhttpVersion}"
     androidTestImplementation "com.squareup.okhttp3:okhttp-tls:${rootProject.ext.okhttpVersion}"
     androidTestUtil 'androidx.test:orchestrator:1.4.1'
